requirements.txt
Pillow==9.2.0
python-dotenv==0.20.0
requests==2.22.0
tweepy==4.10.0
